Fixing schnorr.Verify
