+++ sign/eddsa/eddsa.go
@@ -7,18 +7,12 @@ import (
 	"crypto/sha512"
 	"errors"
 	"fmt"
-	"math/big"
 
 	"go.dedis.ch/kyber/v3"
 	"go.dedis.ch/kyber/v3/group/edwards25519"
 )
 
 var group = new(edwards25519.Curve)
-
-// TODO: maybe export prime and primeOrder from edwards25519/const or allow it to be
-// retrieved from the curve?
-var prime, _ = new(big.Int).SetString("57896044618658097711785492504343953926634992332820282019728792003956564819949", 10)
-var primeOrder, _ = new(big.Int).SetString("7237005577332262213973186563042994240857116359379907606001950938285454250989", 10)
 
 // EdDSA is a structure holding the data necessary to make a series of
 // EdDSA signatures.
@@ -30,15 +24,6 @@ type EdDSA struct {
 
 	seed   []byte
 	prefix []byte
-}
-
-// edDSAPoint is used to verify signatures
-// with checks around canonicality and group order
-type edDSAPoint interface {
-	kyber.Point
-	// HasSmallOrder checks if the given buffer (in little endian)
-	// represents a point with a small order
-	HasSmallOrder() bool
 }
 
 // NewEdDSA will return a freshly generated key pair to use for generating
@@ -143,21 +128,24 @@ func VerifyWithChecks(pub, msg, sig []byte) error {
 	if len(sig) != 64 {
 		return fmt.Errorf("signature length invalid, expect 64 but got %v", len(sig))
 	}
-	if !scalarIsCanonical(sig[32:]) {
+	// The goal of the first comparison is to prevent calling function scalarIsCanonical()
+	// in 99.999% of the cases, saving CPU cycles: this function is called if and only if any
+	// of the most significant 4 bits of sig[32:] are set
+	if (sig[63]&240) > 0 && !edwards25519.ScalarIsCanonical(sig[32:]) {
 		return fmt.Errorf("signature is not canonical")
 	}
-	if !pointIsCanonical(pub) {
+	if !edwards25519.PointIsCanonical(pub) {
 		return fmt.Errorf("public key is not canonical")
 	}
 
-	if !pointIsCanonical(sig[:32]) {
+	if !edwards25519.PointIsCanonical(sig[:32]) {
 		return fmt.Errorf("R is not canonical")
 	}
 	R := group.Point()
 	if err := R.UnmarshalBinary(sig[:32]); err != nil {
 		return fmt.Errorf("got R invalid point: %s", err)
 	}
-	if R.(edDSAPoint).HasSmallOrder() {
+	if R.(kyber.Ed25519Point).HasSmallOrder() {
 		return fmt.Errorf("R has small order")
 	}
 
@@ -170,7 +158,7 @@ func VerifyWithChecks(pub, msg, sig []byte) error {
 	if err := public.UnmarshalBinary(pub); err != nil {
 		return fmt.Errorf("invalid public key: %s", err)
 	}
-	if public.(edDSAPoint).HasSmallOrder() {
+	if public.(kyber.Ed25519Point).HasSmallOrder() {
 		return fmt.Errorf("public key has small order")
 	}
 
@@ -200,60 +188,4 @@ func Verify(public kyber.Point, msg, sig []byte) error {
 		return fmt.Errorf("error unmarshalling public key: %s", err)
 	}
 	return VerifyWithChecks(PBuf, msg, sig)
-}
-
-// scalarIsCanonical whether scalar s is in the range 0<=s<L as required by RFC8032, Section 5.1.7.
-// Also provides Strong Unforgeability under Chosen Message Attacks (SUF-CMA)
-// See paper https://eprint.iacr.org/2020/823.pdf for definitions and theorems
-// See https://github.com/jedisct1/libsodium/blob/4744636721d2e420f8bbe2d563f31b1f5e682229/src/libsodium/crypto_core/ed25519/ref10/ed25519_ref10.c#L2568
-// for a reference
-func scalarIsCanonical(sb []byte) bool {
-	if len(sb) != 32 {
-		return false
-	}
-
-	if sb[31]&0xf0 == 0 {
-		return true
-	}
-
-	L := primeOrder.Bytes()
-	for i, j := 0, 31; i < j; i, j = i+1, j-1 {
-		L[i], L[j] = L[j], L[i]
-	}
-
-	var c byte
-	var n byte = 1
-
-	for i := 31; i >= 0; i-- {
-		// subtraction might lead to an underflow which needs
-		// to be accounted for in the right shift
-		c |= byte((uint16(sb[i])-uint16(L[i]))>>8) & n
-		n &= byte((uint16(sb[i]) ^ uint16(L[i]) - 1) >> 8)
-	}
-
-	return c != 0
-}
-
-// pointIsCanonical determines whether the group element is canonical
-//
-// Checks whether group element s is less than p, according to RFC8032§5.1.3.1
-// https://tools.ietf.org/html/rfc8032#section-5.1.3
-//
-// Taken from
-// https://github.com/jedisct1/libsodium/blob/4744636721d2e420f8bbe2d563f31b1f5e682229/src/libsodium/crypto_core/ed25519/ref10/ed25519_ref10.c#L1113
-func pointIsCanonical(s []byte) bool {
-	if len(s) != 32 {
-		return false
-	}
-
-	c := (s[31] & 0x7f) ^ 0x7f
-	for i := 30; i > 0; i-- {
-		c |= s[i] ^ 0xff
-	}
-
-	// subtraction might underflow
-	c = byte((uint16(c) - 1) >> 8)
-	d := byte((0xed - 1 - uint16(s[0])) >> 8)
-
-	return 1-(c&d&1) == 1
 }
