+++ sign/schnorr/schnorr.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 
 	"go.dedis.ch/kyber/v3"
+	"go.dedis.ch/kyber/v3/group/edwards25519"
 )
 
 // Suite represents the set of functionalities needed by the package schnorr.
@@ -57,9 +58,12 @@ func Sign(s Suite, private kyber.Scalar, msg []byte) ([]byte, error) {
 	return b.Bytes(), nil
 }
 
-// Verify verifies a given Schnorr signature. It returns nil iff the
-// given signature is valid.
-func Verify(g kyber.Group, public kyber.Point, msg, sig []byte) error {
+// VerifyWithChecks uses a public key buffer, a message and a signature.
+// It will return nil if sig is a valid signature for msg created by
+// key public, or an error otherwise. Compared to `Verify`, it performs
+// additional checks around the canonicality and ensures the public key
+// does not have a small order, only when using the ed25519 curve.
+func VerifyWithChecks(g kyber.Group, pub, msg, sig []byte) error {
 	R := g.Point()
 	s := g.Scalar()
 	pointSize := R.MarshalSize()
@@ -68,12 +72,45 @@ func Verify(g kyber.Group, public kyber.Point, msg, sig []byte) error {
 	if len(sig) != sigSize {
 		return fmt.Errorf("schnorr: signature of invalid length %d instead of %d", len(sig), sigSize)
 	}
+
+	if _, is25519 := s.(kyber.Ed25519Scalar); is25519 {
+		if len(sig) != 64 {
+			return fmt.Errorf("signature length invalid, expect 64 but got %v", len(sig))
+		}
+		// The goal of the first comparison is to prevent calling function scalarIsCanonical()
+		// in 99.999% of the cases, saving CPU cycles: this function is called if and only if any
+		// of the most significant 4 bits of sig[32:] are set
+		if (sig[63]&240) > 0 && !edwards25519.ScalarIsCanonical(sig[32:]) {
+			return fmt.Errorf("signature is not canonical")
+		}
+		if !edwards25519.PointIsCanonical(pub) {
+			return fmt.Errorf("public key is not canonical")
+		}
+	}
+
 	if err := R.UnmarshalBinary(sig[:pointSize]); err != nil {
 		return err
 	}
 	if err := s.UnmarshalBinary(sig[pointSize:]); err != nil {
 		return err
 	}
+	public := g.Point()
+	if err := public.UnmarshalBinary(pub); err != nil {
+		return fmt.Errorf("invalid public key: %s", err)
+	}
+
+	if REd25519, is25519 := R.(kyber.Ed25519Point); is25519 {
+		if !edwards25519.PointIsCanonical(sig[:32]) {
+			return fmt.Errorf("R is not canonical")
+		}
+		if REd25519.HasSmallOrder() {
+			return fmt.Errorf("R has small order")
+		}
+		if public.(kyber.Ed25519Point).HasSmallOrder() {
+			return fmt.Errorf("public key has small order")
+		}
+	}
+
 	// recompute hash(public || R || msg)
 	h, err := hash(g, public, R, msg)
 	if err != nil {
@@ -91,6 +128,16 @@ func Verify(g kyber.Group, public kyber.Point, msg, sig []byte) error {
 	}
 
 	return nil
+}
+
+// Verify uses a public key, a message and a Schnorr signature. It will return nil if
+// sig is a valid signature for msg created by key public, or an error otherwise.
+func Verify(g kyber.Group, public kyber.Point, msg, sig []byte) error {
+	PBuf, err := public.MarshalBinary()
+	if err != nil {
+		return fmt.Errorf("error unmarshalling public key: %s", err)
+	}
+	return VerifyWithChecks(g, PBuf, msg, sig)
 }
 
 func hash(g kyber.Group, public, r kyber.Point, msg []byte) (kyber.Scalar, error) {
